requirements.txt
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
einops==0.7.0
matplotlib==3.7.4
mmcv-full==1.5.0
mmengine==0.10.2
mmsegmentation==0.27.0
moviepy==1.0.3
numpy==1.24.3
omegaconf==2.3.0
opencv-python==4.9.0.80
pillow==10.2.0
plyfile==1.0.3
pytorch-ignite==0.4.13
pytorch-lightning==1.3.5
PyWavelets==1.4.1
PyYAML==5.4.1
scikit-image==0.21.0
scipy==1.10.1
tensorboard==2.13.0
tensorboardX==2.6.2.2
timm==0.9.12
torch==1.13.1+cu116
torch-encoding @ git+https://github.com/zhanghang1989/PyTorch-Encoding/@c959dab8312b637fcc7edce83607acb4b0f82645
torchmetrics==0.5.1
torchvision==0.14.1+cu116
tornado==6.4
tqdm==4.66.1
typing_extensions==4.5.0
urllib3==1.26.4